// src/types/supabase.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.15"
  }
  public: {
    Tables: {
// ...
      product_variants: {
        Row: {
          available: boolean | null
          created_at: string | null
          dimensions: string | null
          discount_percentage: number | null
          display_order: number | null
          id: string
          original_price: number
          product_id: string | null
          selling_price: number
          sku: string | null
          stock_quantity: number
          variant_name: string
        }
        Insert: {
          available?: boolean | null
          created_at?: string | null
          dimensions?: string | null
          discount_percentage?: number | null
          display_order?: number | null
          id?: string
          original_price: number
          product_id?: string | null
          selling_price: number
          sku?: string | null
          stock_quantity?: number
          variant_name: string
        }
        Update: {
          available?: boolean | null
          created_at?: string | null
          dimensions?: string | null
          discount_percentage?: number | null
          display_order?: number | null
          id?: string
          original_price?: number
          product_id?: string | null
          selling_price?: number
          sku?: string | null
          stock_quantity?: number
          variant_name?: string
        }
        Relationships: [
          {
            foreignKeyName: "product_variants_product_id_fkey"
            columns: ["product_id"]
            isOneToOne: false
            referencedRelation: "products_dynamic"
            referencedColumns: ["product_id"]
          },
        ]
      }
// ...
      products_dynamic: {
        Row: {
          active: boolean | null
          created_at: string | null
          display_order: number | null
          featured: boolean | null
          new_arrival: boolean | null
          product_id: string
        }
        Insert: {
          active?: boolean | null
          created_at?: string | null
          display_order?: number | null
          featured?: boolean | null
          new_arrival?: boolean | null
          product_id: string
        }
        Update: {
          active?: boolean | null
          created_at?: string | null
          display_order?: number | null
          featured?: boolean | null
          new_arrival?: boolean | null
          product_id?: string
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}